AL - float to int

+++ ordercost.go
@@ -1,6 +1,6 @@
 package main
 
 type OrderCost struct {
-	TotalCost     	float64   	`json:"totalcost"`
-	TotalQuantity  	int   		`json:"totalquantity"`
+	TotalCost     	int   `json:"totalcost"`
+	TotalQuantity  	int   `json:"totalquantity"`
 }
